Ajustes da tradução e cache de imagens

+++ www/js/ctrl/deleteUser.js
@@ -3,17 +3,26 @@ Iris.controller('DeleteUserCtrl', function($scope, $state, $ionicLoading, $ionic
     $scope.searchUser = {};
 
     $scope.deleteUser = function() {
-        var confirmPopup = $ionicPopup.confirm({
-            title: 'Remover',
-            template: 'Deseja realmente remover este usuário?'
-        });
+        if ($scope.searchUser.selected.name == "admin") {
+            $ionicLoading.hide();
+            $ionicPopup.alert({
+                title: 'Falha na remoção.',
+                template: 'Não é possível remover usuário administrador.'
+            });
+        } else {
 
-        confirmPopup.then(function(res) {
-            if (res) {
-                UserService.deleteUser($scope.searchUser.selected.id);
-                $state.go('users');
-            } 
-        });
+            var confirmPopup = $ionicPopup.confirm({
+                title: 'Remover',
+                template: 'Deseja realmente remover este usuário?'
+            });
+
+            confirmPopup.then(function(res) {
+                if (res) {
+                    UserService.deleteUser($scope.searchUser.selected.id);
+                    $state.go('users');
+                }
+            });
+        }
     };
 
     UserService.getUsers().success(function(users) {
